panther_documents/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-currencies @ git+https://github.com/RuVl/django-currencies.git
django-environ==0.11.2
django-recaptcha==3.0.0
idna==3.4
psycopg==3.1.12
psycopg-binary==3.1.12
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
gevent==23.9.1
gunicorn==21.2.0
